# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DealVault API",
    "version": "1.0.0",
    "description": "Commodity deal room management platform API. All endpoints require authentication via NextAuth session unless noted otherwise."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Development" }
  ],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token"
      }
    },
    "schemas": {
...
      "DealStatus": {
        "type": "string",
        "enum": ["draft", "documents_pending", "under_review", "verified", "in_progress", "settled", "closed", "cancelled"]
      },
...
      "Message": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "content": { "type": "string" },
          "visibility": { "type": "string", "enum": ["deal", "side", "private"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "sender": { "$ref": "#/components/schemas/UserSummary" }
        }
      },
...
      "UserSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" }
        }
      },
...
    }
  },
  "security": [{ "session": [] }],
  "paths": {
...
    "/api/deals/{id}": {
      "get": {
        "tags": ["Deals"],
        "summary": "Get deal details with parties, documents, messages, timeline",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Full deal object with nested relations" },
          "403": { "description": "Not a party to this deal" },
          "404": { "description": "Deal not found" }
        }
      },
      "patch": {
        "tags": ["Deals"],
        "summary": "Update deal (status changes enforce state machine)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": { "$ref": "#/components/schemas/DealStatus" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Deal updated" },
          "400": { "description": "Invalid status transition" },
          "403": { "description": "Only creator can update" }
        }
      }
    },
    "/api/deals/{id}/parties": {
      "post": {
        "tags": ["Parties"],
        "summary": "Invite a party to the deal",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "role"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "role": { "type": "string", "enum": ["seller", "buyer", "seller_mandate", "buyer_mandate", "intermediary", "facilitator"] },
                  "commissionPct": { "type": "number", "minimum": 0, "default": 0 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Party invited" },
          "400": { "description": "User not found or already a party" },
          "403": { "description": "Only creator can invite" }
        }
      }
    },
    "/api/deals/{id}/parties/{partyId}": {
      "patch": {
        "tags": ["Parties"],
        "summary": "Accept or reject party invitation",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "partyId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["action"],
                "properties": {
                  "action": { "type": "string", "enum": ["accept", "reject"] },
                  "companyId": { "type": "string", "description": "Optional company to link" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Party status updated" },
          "403": { "description": "Can only respond to own invitation" }
        }
      }
    },
    "/api/deals/{id}/messages": {
      "get": {
        "tags": ["Messages"],
        "summary": "List messages (filtered by visibility)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Array of messages", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Message" } } } } }
        }
      },
      "post": {
        "tags": ["Messages"],
        "summary": "Send a message in the deal room",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["content"],
                "properties": {
                  "content": { "type": "string" },
                  "visibility": { "type": "string", "enum": ["deal", "side", "private"], "default": "deal" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Message sent" }
        }
      }
    },
...
  }
}